+++ src/main/kotlin/minesweeper/domain/Mine.kt
@@ -1,3 +1,3 @@
-package minesweeper
+package minesweeper.domain
 
 class Mine (val position: Position){}
