+++ src/main/kotlin/minesweeper/domain/Position.kt
@@ -1,3 +1,3 @@
-package minesweeper
+package minesweeper.domain
 
 data class Position(val width: Int, val height: Int)
