+++ src/test/kotlin/minesweeper/domain/MineTest.kt
@@ -1,6 +1,8 @@
-package minesweeper
+package minesweeper.domain
 
 import io.kotest.matchers.shouldBe
+import minesweeper.domain.Mine
+import minesweeper.domain.Position
 import org.junit.jupiter.api.Test
